+++ utils/devices.ts
@@ -1,0 +1,11 @@
+const sizes = {
+  tablet: '768px',
+  laptop: '1440px',
+}
+
+const devices = {
+  tablet: `min-width: ${sizes.tablet}`,
+  laptop: `min-width: ${sizes.laptop}`,
+}
+
+export { devices }
